+++ test/tilt_template_test.rb
@@ -142,6 +142,53 @@ class TiltTemplateTest < MiniTest::Unit::TestCase
     assert_equal "1 + 2 = 3", inst.render(Object.new, '_answer' => 3)
   end
 
+  class CustomGeneratingMockTemplate < PreparingMockTemplate
+    def precompiled_template(locals)
+      data
+    end
+
+    def precompiled_preamble(locals)
+      options.fetch(:preamble)
+    end
+
+    def precompiled_postamble(locals)
+      options.fetch(:postamble)
+    end
+  end
+
+  test "supports pre/postamble" do
+    inst = CustomGeneratingMockTemplate.new(
+      :preamble => 'buf = []',
+      :postamble => 'buf.join'
+    ) { 'buf << 1' }
+
+    assert_equal "1", inst.render
+  end
+
+  # Special-case for Haml
+  # https://github.com/rtomayko/tilt/issues/193
+  test "supports Array pre/postambles" do
+    inst = CustomGeneratingMockTemplate.new(
+      :preamble => ['buf = ', '[]'],
+      :postamble => ['buf.', 'join']
+    ) { 'buf << 1' }
+
+    # TODO: Use assert_output when we swicth to MiniTest
+    warns = <<-EOF
+TiltTemplateTest::CustomGeneratingMockTemplate#precompiled_preamble should return String (not Array)
+TiltTemplateTest::CustomGeneratingMockTemplate#precompiled_postamble should return String (not Array)
+EOF
+
+    begin
+      require 'stringio'
+      $stderr = StringIO.new
+      assert_equal "1", inst.render
+      assert_equal warns, $stderr.string
+    ensure
+      $stderr = STDERR
+    end
+  end
+
   class Person
     CONSTANT = "Bob"