+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## master
 
+* Make template compilation thread-safe (#191, judofyr)
+
+## 1.4.1 (2013-05-08)
+
+* Support Arrays in pre/postambles (#193, jbwiv)
+
 ## 1.4.0 (2013-05-01)
 
 * Better encoding support